requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Cors==4.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
pytest==7.4.0
gunicorn==21.2.0 